build: use allowed over enabled or not-disabled
Meson feature options are typically in a tri-state of enabled, disabled,
or auto.  The enabled and disabled functions on an option (from
`get_option`) no longer return true for auto features.  Instead, the
expectation is to use `allowed()` which is true for both enabled and auto.

Switch all uses of `enabled` to `allowed`.
Switch all uses of `not disabled` to `allowed`.

Change-Id: I9ccd43d77594fb735df9d74d78cb42da8bdc0685
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>